requirements.txt
Django==3.2.0
djangorestframework==3.13.1
psycopg2
python-decouple